ai_engine/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0
openai==0.28.1
python-dotenv==1.0.0
google-generativeai==0.4.1
